refactor(LottoResult): totalPrize double로 수정

+++ src/main/java/lotto/domain/LottoResult.java
@@ -45,7 +45,7 @@ public class LottoResult {
     }
 
     public double calculateRateOfReturn(int amount) {
-        int totalPrize = calculateTotalPrize();
+        double totalPrize = calculateTotalPrize();
         return (totalPrize / amount) * 100;
     }
 }
